refactor(stelae): move MithrilFeedback to the binary's feedback module
The daemon extraction had widened it to pub(crate) so snapshot backfill
could reach into bootstrap mithril's module for a TUI adapter — a
command reaching into another command. The receiver both mithril-driven
commands build belongs in feedback.rs beside the binary's other
renderers, private to nothing but the feature gate.
